refactor : 공통 PageResponse 형식 코드에 반영 (#108)
