--- ourhome-market/src/main/java/com/armand/ourhome/market/voucher/dto/PageResponse.java
@@ -1,20 +1,0 @@
-package com.armand.ourhome.market.voucher.dto;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-@Builder
-@JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class PageResponse<T> {
-
-	private T content;
-	private long totalElements;
-	private int totalPages;
-	private int size;
-
-}
